requirements.txt
# Core
python-dotenv>=1.0.0

# LangChain & LLM
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0

# Embeddings
sentence-transformers>=2.6.0
langchain-huggingface>=0.0.3

# Vector Database
chromadb>=0.5.0

# Utilities
tqdm>=4.66.0
numpy>=1.24.0
